requirements.txt
numpy==2.2.0
opencv-python==4.11.0.86
mediapipe==0.10.21
pillow==11.1.0